requirements.txt
###talib

#tar -xvf ta-lib-0.4.0-src.tar.gz
#cd ta-lib
#./configure --prefix=/usr
#make
#sudo make install
#sudo ldconfig



python-binance
TA-Lib
numpy
websocket_client
